requirements.txt
asgiref==3.3.1
click==7.1.2
dbconfig==0.3.2
Django==3.1.4
Flask==1.1.2
Flask-Cors==3.0.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.22
protobuf==3.14.0
pyclean==2.0.0
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
Werkzeug==1.0.1
